pyproject.toml
[project]
name = "cybercardash"
version = "1.0.0"
description = "Industrial-grade General Electronic Competition Control Host Computer"
authors = [
    {name = "Trae Solo", email = "trae@example.com"},
]
requires-python = ">=3.11"
dependencies = [
    "PySide6>=6.6.0",
    "pyqtgraph>=0.13.3",
    "pyserial>=3.5",
    "numpy>=1.26.0",
    "pydantic>=2.5.0",
    "scipy>=1.11.0"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
